chennai-bootcamp-playground/requirements.txt
autopep8==2.0.1
bcrypt==4.0.1
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.1
Werkzeug==2.2.3
